apps/backend/requirements.txt
fastapi>=0.109
uvicorn[standard]>=0.27
pymysql>=1.1.0
python-dotenv>=1.0
python-jose[cryptography]>=3.3
google-auth>=2.28
anthropic>=0.28
apscheduler>=3.10
httpx>=0.27
fpdf2>=2.7
pytesseract>=0.3.10
pdf2image>=1.17
Pillow>=10.0
selenium>=4.15
webdriver-manager>=4.0
python-multipart>=0.0.9
pytest>=7.4
pytest-asyncio>=0.23
openai>=1.0
cryptography>=42.0
google-genai>=0.8
google-ads>=24.0.0
rembg>=2.0.50
